Make list item generic component and move location
