+++ src/components/SimpleListItem/index.js
@@ -1,0 +1,3 @@
+import SimpleListItem from './SimpleListItem';
+
+export default SimpleListItem;
