--- src/views/UnitView/styles/listStyles.js
@@ -1,38 +1,0 @@
-export default () => ({
-  title: {
-    float: 'left',
-    marginLeft: '15px',
-    marginTop: '12px',
-    marginBottom: '12px',
-    fontWeight: 700,
-    fontSize: '1.030em',
-    lineHeight: 1.5,
-  },
-  listItem: {
-    height: '56px',
-    paddingLeft: 0,
-    paddingRight: 0,
-    paddingTop: '7px',
-    paddingBottom: '9px',
-  },
-  textContainer: {
-    padding: 0,
-    marginLeft: '16px',
-    marginRight: '42px',
-  },
-  listLinkText: {
-    fontWeight: 400,
-    fontSize: '0.911em',
-  },
-  divider: {
-    marginLeft: '72px',
-  },
-  listIcon: {
-    width: '24px',
-    height: '24px',
-    margin: '16px',
-  },
-  link: {
-    color: '#0000EE',
-  },
-});
